vdin: Fixed screencatch screenshot restart issue [1/1]
PD#SWPL-179076

Problem:
420 output, the C address configured to the register
is the wrong data type, causing memory to be flushed

Solution:
Example Change the data type of address C to the correct one

Verify:
s6

+++ drivers/media/vin/tvin/vdin/vdin_drv.c
@@ -971,7 +971,7 @@ static void vdin_vf_init(struct vdin_dev_s *devp)
 					devp->v4l2_fmt.fmt.pix_mp.num_planes == 2)
 					phy_c_addr = devp->vf_mem_c_start[i];
 				else
-					phy_c_addr = (u32)(devp->vf_mem_start[i] + luma_size);
+					phy_c_addr = devp->vf_mem_start[i] + luma_size;
 				vf->canvas0_config[1].phy_addr = phy_c_addr;
 				vf->canvas0_config[1].width = devp->canvas_w;
 				vf->canvas0_config[1].height =
